мини фикс int to decimal

+++ Model/MeasurementsData.cs
@@ -2,9 +2,9 @@
 {
     public class MeasurementsData
     {
-        public int Temperature { get; set; }
-        public int Humidity { get; set; }
-        public int Pressure { get; set; }
+        public decimal Temperature { get; set; }
+        public decimal Humidity { get; set; }
+        public decimal Pressure { get; set; }
         public TypeVerification TypeVerification { get; set; }
         /// <summary>
         /// Поверитель , в месте использования будет сделана коллекция имен или типо того
